finished math expression interpreter

+++ src/lexer.rs
@@ -7,10 +7,10 @@ pub fn tokenize(input_string: &str) -> Result<Vec<Token>, String> {
     while let Some(&c) = chars.peek() {
         match c {
             '0'..='9' => {
-                let mut n: u32 = 0;
+                let mut n: f64 = 0.;
                 while let Some(&i) = chars.peek() {
                     if i.is_digit(10) {
-                        n = n * 10 + i.to_digit(10).unwrap();
+                        n = n * 10. + i.to_digit(10).unwrap() as f64;
                         chars.next();
                     }
                     else {
